+++ src/logChecker/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "4.2.2"
-__author__ = 'Lucas Aimaretto, Beatriz Bonafe, Kathleen Mendonca'
+__version__ = "4.5.6"
+__author__ = 'Lucas Aimaretto, Beatriz Bonafe'